% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.jointmeta1SE}
\alias{confint.jointmeta1SE}
\title{Extract confidence intervals}
\usage{
\method{confint}{jointmeta1SE}(fittedSE)
}
\arguments{
\item{fittedSE}{A jointmeta1SE object - the result of running the
bootstrapping function \code{jointmetaSE} on a jointmeta1 object}
}
\value{
Returns the name of variables, the part of the joint model they
    relate to (sub-model, variance parameter...), their estimate and
    their 95\% confidence interval
}
\description{
\code{confint} returns the bootstrapped confidence intervals from a
    \code{jointmeta1SE} object, which holds the results of bootstrapping the
    fit from a \code{jointmeta1} function fit.
}
\examples{
    #change data to jointdata format
    jointdat<-tojointdata(longitudinal = simdat$longitudinal,
                          survival = simdat$survival, id = "id",
                          longoutcome = "Y", timevarying = c("time","ltime"),
                          survtime = "survtime", cens = "cens",
                          time = "time")

    #ensure variables are correctly formatted
    jointdat$baseline$study <- as.factor(jointdat$baseline$study)
    jointdat$baseline$treat <- as.factor(jointdat$baseline$treat)

    #fit multi-study joint model and calculate the standard errors
    onestagefit<-jointmeta1(data = jointdat, long.formula = Y ~ 1 + time +
                            treat + study, long.rand.ind = c("int", "time"),
                            long.rand.stud = c("treat"),
                            sharingstrct = "randprop",
                            surv.formula = Surv(survtime, cens) ~ treat,
                            study.name = "study", strat = T)

    onestagefitSE<-jointmetaSE(fitted = onestagefit, n.boot = 200)

    #extract confidence intervals
    confint(onestagefitSE)

}
\seealso{
\code{\link{jointmeta1}}
}
