% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.jointmeta1.R
\name{print.jointmeta1}
\alias{print.jointmeta1}
\title{Print function for \code{jointmeta1} objects}
\usage{
\method{print}{jointmeta1}(fitted)
}
\arguments{
\item{fitted}{a jointmeta1.object, the result of fitting a jointmeta1 model}
}
\value{
An object inheriting from class \code{print.jointmeta1} with all
  components included in \code{fitted} (see \code{\link{jointmeta1.object}}).
}
\description{
A function to print a \code{\link{jointmeta1.object}}.
}
\examples{
    #change data to jointdata format
    jointdat<-tojointdata(longitudinal = simdat$longitudinal,
                          survival = simdat$survival, id = 'id',
                          longoutcome = 'Y', timevarying = c('time','ltime'),
                          survtime = 'survtime', cens = 'cens',
                          time = 'time')

    #ensure variables are correctly formatted
    jointdat$baseline$study <- as.factor(jointdat$baseline$study)
    jointdat$baseline$treat <- as.factor(jointdat$baseline$treat)

    #fit multi-study joint model
    onestagefit<-jointmeta1(data = jointdat, long.formula = Y ~ 1 + time +
                            treat + study, long.rand.ind = c('int', 'time'),
                            long.rand.stud = c('treat'),
                            sharingstrct = 'randprop',
                            surv.formula = Surv(survtime, cens) ~ treat,
                            study.name = 'study', strat = T)

    #print the fitted multi-study joint model
    print(onestagefit)

}
