% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.jointmeta1}
\alias{summary.jointmeta1}
\title{Summary function for jointmeta1}
\usage{
\method{summary}{jointmeta1}(fitted, variance = TRUE)
}
\arguments{
\item{fitted}{a jointmeta1.object, the result of fitting a jointmeta1 model}

\item{variance}{a logical if set to \code{TRUE} the variances of the
measurement errors and the random effects are returned, else if
\code{FALSE} then the standard errors are returned.}
}
\value{
An object inheriting from class \code{summary.jointmeta1} with all
    components included in \code{fitted} (see
    \code{\link{jointmeta1.object}}).
}
\description{
A function to provide a summary of a
    \code{\link{jointmeta1.object}}.
}
\examples{
    #change data to jointdata format
    jointdat<-tojointdata(longitudinal = simdat$longitudinal,
                          survival = simdat$survival, id = "id",
                          longoutcome = "Y", timevarying = c("time","ltime"),
                          survtime = "survtime", cens = "cens",
                          time = "time")

    #ensure variables are correctly formatted
    jointdat$baseline$study <- as.factor(jointdat$baseline$study)
    jointdat$baseline$treat <- as.factor(jointdat$baseline$treat)

    #fit multi-study joint model
    onestagefit<-jointmeta1(data = jointdat, long.formula = Y ~ 1 + time +
                            treat + study, long.rand.ind = c("int", "time"),
                            long.rand.stud = c("treat"),
                            sharingstrct = "randprop",
                            surv.formula = Surv(survtime, cens) ~ treat,
                            study.name = "study", strat = T)

    #request a summary of the fitted model, with variances printed
    summary(onestagefit, variance = TRUE)

    #request a summary of the fitted model, with standard errors printed
    summary(onestagefit, variance = FALSE)
}
