% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.jointmeta1SE}
\alias{vcov.jointmeta1SE}
\title{Extract the variance covariance matrix from the bootstrapped results}
\usage{
\method{vcov}{jointmeta1SE}(fittedSE)
}
\arguments{
\item{fittedSE}{an object of class \code{jointmeta1SE}}
}
\value{
a variance covariance matrix for the fixed effects from the
  longitudinal sub-model, the time-to-event sub-model, the association
  parameters, the random effects and the error term.
}
\description{
Function applied to a \code{jointmeta1SE} object, the result of the
\code{jointmetaSE} function to extract the variance covariance matrix for the
estimated model parameters
}
\examples{

    #change data to jointdata format
    jointdat<-tojointdata(longitudinal = simdat$longitudinal,
                          survival = simdat$survival, id = 'id',
                          longoutcome = 'Y', timevarying = c('time','ltime'),
                          survtime = 'survtime', cens = 'cens',time = 'time')

    #ensure variables are correctly formatted
    jointdat$baseline$study <- as.factor(jointdat$baseline$study)
    jointdat$baseline$treat <- as.factor(jointdat$baseline$treat)

    #example without interaction terms
    #fit multi-study joint model
    onestagefit<-jointmeta1(data = jointdat, long.formula = Y ~ 1 + time +
                            treat + study, long.rand.ind = c('int', 'time'),
                            long.rand.stud = c('treat'),
                            sharingstrct = 'randprop',
                            surv.formula = Surv(survtime, cens) ~ treat,
                            study.name = 'study', strat = T)

    #calculate the SE
    onestagefitSE <- jointmetaSE(fitted = onestagefit, n.boot = 200)

    #extract the variance covariance matrix
    vcov(onestagefitSE)

    #example with interaction terms
    #fit multi-study joint model
    onestagefit2<-jointmeta1(data = jointdat, long.formula = Y ~ 1 + time +
                           treat*study, long.rand.ind = c('int', 'time'),
                           sharingstrct = 'randprop',
                           surv.formula = Surv(survtime, cens) ~ treat*study,
                           study.name = 'study', strat = F)

    #calculate the SE
    onestagefit2SE <- jointmetaSE(fitted = onestagefit2, n.boot = 125,
          overalleffects = list(long = list(c('treat1', 'treat1:study2'),
                                            c('treat1', 'treat1:study3'),
                                            c('treat1', 'treat1:study4'),
                                            c('treat1', 'treat1:study5')),
                                surv = list(c('treat1', 'treat1:study2'),
                                            c('treat1', 'treat1:study3'),
                                            c('treat1', 'treat1:study4'),
                                            c('treat1', 'treat1:study5'))))

}
\seealso{
\code{\link{jointmeta1}}, \code{\link{jointmeta1SE}},
  \code{\link{jointmeta1SE.object}}
}
